Summarise the proportions of a df (population) which fall in each of n categories

produce a summary table showing the proportion of a numeric column within groups and subgroups, which are assigned by the values in the corresponding row of other columns

prop_in_group(
  input_df,
  breakdowns_vector,
  value_col = NULL,
  group_by_col = NULL,
  value_or_prop = "prop",
  prop_dps = 2,
  show_knowns_cols = "yes",
  knowns_treatment = "count",
  round_knowns_to_nearest = 1
)

Arguments

input_df

an input data frame

breakdowns_vector

a vector containing the column names of each group to see the proportion for

value_col

the population values to be summed. If the purpose is to count rows, leave this blank blank.

group_by_col

an additional layer of grouping, if required, to see a broader subset

value_or_prop

default = "prop". Inputting "value" can generate a df with values instead of proportions

prop_dps

how many decimal places to make the proportion. The default is 2, representing a percentage with no decimals

show_knowns_cols

whether to show the cols with count or sum of NA values in the df. Default = "yes"

knowns_treatment

inlcude "sum" or "count" to specify how to calculate the unknowns/known instances column

round_knowns_to_nearest

the number to round the knowns/unknowns to e.g. 1,5,10

Value

The output from the function - a data frame with 5+ columns: grouping (containing each element in the breakdowns vector); subgroup (containing each unique value within each grouping); a column per group_by_col unique value, if none, this will just be the prop_in_total_group; how many rows had known values; and how many unknowns there were

See also